+++ rhoc/pkg/cmd/connectors/connectors.go
@@ -1,10 +1,10 @@
 package connectors
 
 import (
-	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/cmd/commands"
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/cmd/connectors/delete"
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/cmd/connectors/describe"
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/cmd/connectors/list"
+	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/commands"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
 	"github.com/spf13/cobra"
 )
