+++ rhoc/pkg/cmd/namespaces/delete/delete.go
@@ -1,7 +1,7 @@
 package delete
 
 import (
-	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/cmd/commands"
+	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/commands"
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/service"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/dump"
